requirements.txt
Django==3.2.13
psycopg2-binary
#sqlparse
djangorestframework==3.11.1